Split CLI build graph execution
